before combining into one

+++ node-delete/delete.node.mjs
@@ -1,25 +1,51 @@
 import fs from 'fs';
-import { display, displayError, displayInTheMiddle } from '@createdbyjurand/node-display';
+import {display, displayError, displayInTheMiddle} from '@createdbyjurand/node-display';
+import {node} from '@createdbyjurand/node-node';
+import {prefix} from '@createdbyjurand/prefix';
 
 displayInTheMiddle('delete.node.mjs version 0.4.0');
 
 export const deleteFile = fileName => {
-  display(fileName + ': Deleting', '[   OK   ]');
+  display(`${prefix(fileName)} Deleting`, '[   OK   ]');
   if (fs.existsSync(fileName)) {
     fs.unlinkSync(fileName);
-    display(fileName + ': Deleted', '[   OK   ]');
+    display(`${prefix(fileName)} Deleted`, '[   OK   ]');
   } else {
-    displayError(fileName + ': File not found', '[ FAILED ]');
+    displayError(`${prefix(fileName)} File not found`, '[ FAILED ]');
   }
 };
 
 export const deleteFolder = folderName => {
-  display(folderName + ': Deleting', '[   OK   ]');
-  if (fs.existsSync(folderName)) {
-    fs.rmSync(folderName, { recursive: true });
-    display(folderName + ': Deleted', '[   OK   ]');
+  if (fs.existsSync(folderName) && !folderName.startsWith('.') && !folderName.startsWith('/')) {
+    if (node.version >= 14) {
+      display(`${prefix(folderName)} Detecting Node version...`, '[   OK   ]');
+      display(`${prefix(folderName)} Node ${node.version} detected`, '[   OK   ]');
+      display(`${prefix(folderName)} Deleting using rmSync`, '[   OK   ]');
+      fs.rmSync(folderName, {recursive: true});
+    } else if (node.version >= 12) {
+      display(`${prefix(folderName)} Detecting Node version...`, '[   OK   ]');
+      display(`${prefix(folderName)} Node ${node.version} detected`, '[   OK   ]');
+      display(`${prefix(folderName)} Deleting using rmdirSync`, '[   OK   ]');
+      fs.rmdirSync(folderName, {recursive: true});
+    } else {
+      display(`${prefix(folderName)} Detecting Node version...`, '[   OK   ]');
+      display(`${prefix(folderName)} Node ${node.version} detected`, '[   OK   ]');
+      display(`${prefix(folderName)} Deleting manually...`, '[   OK   ]');
+      fs.readdirSync(folderName).forEach(file => {
+        const curPath = `${folderName}/${file}`;
+        if (fs.lstatSync(curPath).isDirectory()) {
+          // recurse
+          deleteFolder(curPath);
+        } else {
+          // delete file
+          fs.unlinkSync(curPath);
+        }
+      });
+      fs.rmdirSync(folderName);
+    }
+    display(`${prefix(folderName)} Deleted`, '[   OK   ]');
   } else {
-    displayError(folderName + ': Folder not found', '[ FAILED ]');
+    displayError(`${prefix(folderName)} Folder not found`, '[ FAILED ]');
   }
 };
 
@@ -29,7 +55,7 @@ export const deleteFolderExcept = (folderName, except) => {
     fs.readdirSync(folderName).forEach(subfolder => {
       const curPath = path.join(folderName, subfolder);
       if (fs.lstatSync(curPath).isDirectory()) {
-        fs.rmdirSync(curPath, { recursive: true });
+        fs.rmdirSync(curPath, {recursive: true});
       } else {
         if (shouldDelete(subfolder)) {
           fs.unlinkSync(curPath);
